+++ backend/cwa_lib/app.py
@@ -2,15 +2,14 @@ from contextlib import asynccontextmanager
 import os
 from traceback import format_exc
 
-from fastapi import FastAPI, Request
+from fastapi import FastAPI, Request, Depends
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.staticfiles import StaticFiles
 from fastapi.templating import Jinja2Templates
 
 from common import FRONTEND_DIR_STATIC, CORS_ORIGINS, log
 
-from cwa_lib.users import fastapi_users
-from fastapi_users.password import PasswordHelper
+from cwa_lib.middleware.last_seen import refresh_last_seen
 
 
 @asynccontextmanager
@@ -20,7 +19,12 @@ async def lifespan(app: FastAPI):
     # Clean up after web app stop
     pass
 
-app = FastAPI(title="Cognosa Tasks API", version="1.0", lifespan=lifespan)
+app = FastAPI(
+    title="Cognosa Tasks API",
+    version="1.0",
+    lifespan=lifespan,
+    dependencies=[Depends(refresh_last_seen)],
+)
 
 app.add_middleware(
     CORSMiddleware,
@@ -34,10 +38,6 @@ app.mount('/static', StaticFiles(directory=FRONTEND_DIR_STATIC), name="static")
 app.mount("/assets", StaticFiles(directory=os.path.join(FRONTEND_DIR_STATIC, "assets")), name="assets")
 templates = Jinja2Templates(directory=FRONTEND_DIR_STATIC)
 
-# A tiny protected endpoint (alt to /users/me)
-current_active_user = fastapi_users.current_user(active=True)
-current_active_user_or_none = fastapi_users.current_user(active=True, optional=True)
-password_helper = PasswordHelper()
 
 
 def get_client_ip(request: Request) -> str: