+++ backend/cwa_lib/routers/misc.py
@@ -4,7 +4,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from common.sql_db_async import async_get_session
 from common.sql_models import User
 from cwa_lib.pydantic_schemas.user import ChangePasswordRequest
-from cwa_lib.app import current_active_user, password_helper
+from cwa_lib.users import current_active_user, password_helper
 from cwa_lib.pages.server_status import ServerStatusPage