+++ backend/cwa_lib/routers/su_change_oneself.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Depends, HTTPException
 from common.sql_db_async import AsyncSession, async_get_session
 from common.sql_models.api_users import User
-from cwa_lib.app import current_active_user
+from cwa_lib.users import current_active_user
 from cwa_lib.pydantic_schemas.su_change_oneself import (
     SuChangeOneselfGetResult, SuChangeOneselfUpdate, SuChangeOneselfUpdateResult
 )