+++ backend/cwa_lib/routers/su_manage_api_settings.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Depends, Request, HTTPException
 from common.sql_db_async import AsyncSession, async_get_session
 from common.sql_models.api_users import User
-from cwa_lib.app import current_active_user
+from cwa_lib.users import current_active_user
 from cwa_lib.pydantic_schemas.generic_table import TableQuery, TableCreateRowResult, TableUpdateRowResult
 from cwa_lib.pydantic_schemas.su_manage_api_settings import (
     SuManageApiSettingsQueryResult, SuManageApiSettingsCreate, SuManageApiSettingsUpdate