+++ backend/cwa_lib/routers/su_server_status.py
@@ -3,7 +3,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from common.sql_db_async import async_get_session
 from common.sql_models import User
-from cwa_lib.app import current_active_user
+from cwa_lib.users import current_active_user
 from cwa_lib.pages.su_server_status import SuServerStatusPage