+++ backend/cwa_lib/routers/users.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Depends, HTTPException, status
 from common.sql_db_async import AsyncSession, async_get_session
-from common.sql_models import User, ApiGroups
-from cwa_lib.app import current_active_user
+from common.sql_models import User
+from cwa_lib.users import current_active_user
 from cwa_lib.pydantic_schemas.user import UsersMe
 from cwa_lib.sql_tables.api_groups import ApiGroupsTable